new eclipse warning: remove unnecessary cast.
git-svn-id: https://geobeagle.googlecode.com/svn@665 eab1e41a-c3b7-11dd-b16b-519e865567fa

+++ branches/sng/GeoBeagle/src/com/google/code/geobeagle/xmlimport/CacheTagWriter.java
@@ -50,7 +50,7 @@ public class CacheTagWriter {
 
         public int stars(String stars) {
             try {
-                return (int)Math.round(Float.parseFloat(stars) * 2);
+                return Math.round(Float.parseFloat(stars) * 2);
             } catch (Exception ex) {
                 return 0;
             }
